Fixed favorite routes and created authentication verification middleware

+++ src/modules/favorites/services/DeleteFavoriteService.ts
@@ -1,23 +1,24 @@
 import { inject, injectable } from 'tsyringe';
 import { DeleteResult } from 'typeorm';
 
+import AppError from '@app/errors/AppError';
+
 import FavoritesRepository from '../repositories/FavoritesRepository';
 
-interface Request {
-  userId: number;
-  drinkId: number;
-}
-
 @injectable()
-export default class DeleteUserService {
+export default class DeleteFavoriteService {
   constructor(
     @inject('FavoritesRepository')
     private favoritesRepository: FavoritesRepository,
   ) {}
 
-  public async execute({ userId, drinkId }: Request): Promise<DeleteResult> {
-    const user = await this.favoritesRepository.delete({ userId, drinkId });
+  public async execute(userId: number, drinkId: number): Promise<DeleteResult> {
+    if (!userId || !drinkId) {
+      throw new AppError('Você deve preencher todos os campos');
+    }
 
-    return user;
+    const favorite = await this.favoritesRepository.delete({ userId, drinkId });
+
+    return favorite;
   }
 }
